+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - tRNA coordinates are now 1-indexed.
+- Write `summary_execution_info`.
 
 ## [1.0.0] - 2022-07-12
 ### Added